--- NMPopUpView/AppDelegate.h
@@ -1,15 +1,0 @@
-//
-//  AppDelegate.h
-//  NMPopUpView
-//
-//  Created by Nikos Maounis on 9/12/13.
-//  Copyright (c) 2013 Nikos Maounis. All rights reserved.
-//
-
-#import <UIKit/UIKit.h>
-
-@interface AppDelegate : UIResponder <UIApplicationDelegate>
-
-@property (strong, nonatomic) UIWindow *window;
-
-@end
