--- NMPopUpView/Base.lproj/Main.storyboard
@@ -1,46 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" standalone="no"?>
-<document type="com.apple.InterfaceBuilder3.CocoaTouch.Storyboard.XIB" version="3.0" toolsVersion="6254" systemVersion="14B25" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" initialViewController="vXZ-lx-hvc">
-    <dependencies>
-        <deployment identifier="iOS"/>
-        <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="6247"/>
-    </dependencies>
-    <scenes>
-        <!--Demo View Controller-->
-        <scene sceneID="ufC-wZ-h7g">
-            <objects>
-                <viewController id="vXZ-lx-hvc" customClass="DemoViewController" customModule="NMPopUpView" sceneMemberID="viewController">
-                    <layoutGuides>
-                        <viewControllerLayoutGuide type="top" id="iIO-T4-jd8"/>
-                        <viewControllerLayoutGuide type="bottom" id="isS-RT-QZM"/>
-                    </layoutGuides>
-                    <view key="view" contentMode="scaleToFill" id="kh9-bI-dsS">
-                        <rect key="frame" x="0.0" y="0.0" width="600" height="600"/>
-                        <autoresizingMask key="autoresizingMask" flexibleMaxX="YES" flexibleMaxY="YES"/>
-                        <subviews>
-                            <button opaque="NO" contentMode="scaleToFill" contentHorizontalAlignment="center" contentVerticalAlignment="center" buttonType="roundedRect" lineBreakMode="middleTruncation" translatesAutoresizingMaskIntoConstraints="NO" id="WhB-eO-Xy8">
-                                <rect key="frame" x="87" y="286" width="426" height="30"/>
-                                <state key="normal" title="Show Popup">
-                                    <color key="titleShadowColor" white="0.5" alpha="1" colorSpace="calibratedWhite"/>
-                                </state>
-                                <connections>
-                                    <action selector="showPopUp:" destination="vXZ-lx-hvc" eventType="touchUpInside" id="FSH-f7-bsV"/>
-                                </connections>
-                            </button>
-                        </subviews>
-                        <color key="backgroundColor" white="1" alpha="1" colorSpace="custom" customColorSpace="calibratedWhite"/>
-                        <constraints>
-                            <constraint firstItem="WhB-eO-Xy8" firstAttribute="centerY" secondItem="kh9-bI-dsS" secondAttribute="centerY" constant="1" id="71t-T9-cWP"/>
-                            <constraint firstItem="WhB-eO-Xy8" firstAttribute="centerX" secondItem="kh9-bI-dsS" secondAttribute="centerX" id="Wtb-Hb-BeK"/>
-                            <constraint firstItem="WhB-eO-Xy8" firstAttribute="leading" secondItem="kh9-bI-dsS" secondAttribute="leading" constant="87" id="Wtj-tY-Tt2"/>
-                        </constraints>
-                    </view>
-                    <connections>
-                        <outlet property="showPopupBtn" destination="WhB-eO-Xy8" id="V3W-T6-mAa"/>
-                    </connections>
-                </viewController>
-                <placeholder placeholderIdentifier="IBFirstResponder" id="x5A-6p-PRh" sceneMemberID="firstResponder"/>
-            </objects>
-            <point key="canvasLocation" x="654.375" y="366.54929577464787"/>
-        </scene>
-    </scenes>
-</document>
