+++ NMPopUpView/Example/Podfile
@@ -1,0 +1,12 @@
+source 'https://github.com/CocoaPods/Specs.git'
+use_frameworks!
+
+target 'NMPopUpViewSwift_Example', :exclusive => true do
+  pod "NMPopUpViewSwift", :path => "../"
+end
+
+target 'NMPopUpViewSwift_Tests', :exclusive => true do
+  pod "NMPopUpViewSwift", :path => "../"
+
+  
+end
