+++ NMPopUpView/Pod/Classes/PopUpViewController_iPad.xib
@@ -1,0 +1,77 @@
+<?xml version="1.0" encoding="UTF-8" standalone="no"?>
+<document type="com.apple.InterfaceBuilder3.CocoaTouch.iPad.XIB" version="3.0" toolsVersion="8191" systemVersion="14F27" targetRuntime="iOS.CocoaTouch.iPad" propertyAccessControl="none" useAutolayout="YES">
+    <dependencies>
+        <deployment identifier="iOS"/>
+        <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="8154"/>
+    </dependencies>
+    <objects>
+        <placeholder placeholderIdentifier="IBFilesOwner" id="-1" userLabel="File's Owner" customClass="PopUpViewController">
+            <connections>
+                <outlet property="logoImg" destination="gmN-gy-bjy" id="QgK-YD-PTP"/>
+                <outlet property="messageLabel" destination="zzP-bh-DHM" id="Vcd-7j-vWE"/>
+                <outlet property="popUpView" destination="cYT-D9-Z5J" id="idF-M2-Ksa"/>
+                <outlet property="view" destination="1" id="b8x-g2-m73"/>
+            </connections>
+        </placeholder>
+        <placeholder placeholderIdentifier="IBFirstResponder" id="-2" customClass="UIResponder"/>
+        <view contentMode="scaleToFill" id="1">
+            <rect key="frame" x="0.0" y="0.0" width="768" height="1024"/>
+            <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
+            <subviews>
+                <view contentMode="scaleToFill" translatesAutoresizingMaskIntoConstraints="NO" id="cYT-D9-Z5J">
+                    <rect key="frame" x="190" y="287" width="389" height="449"/>
+                    <subviews>
+                        <button opaque="NO" contentMode="scaleToFill" contentHorizontalAlignment="center" contentVerticalAlignment="center" buttonType="roundedRect" lineBreakMode="middleTruncation" translatesAutoresizingMaskIntoConstraints="NO" id="Ple-bp-UYe">
+                            <rect key="frame" x="111" y="379" width="166" height="30"/>
+                            <constraints>
+                                <constraint firstAttribute="height" constant="30" id="AVp-AL-fTV"/>
+                                <constraint firstAttribute="width" constant="166" id="psK-Xs-V1a"/>
+                            </constraints>
+                            <fontDescription key="fontDescription" type="boldSystem" pointSize="15"/>
+                            <state key="normal" title="Close">
+                                <color key="titleShadowColor" white="0.5" alpha="1" colorSpace="calibratedWhite"/>
+                            </state>
+                            <connections>
+                                <action selector="closePopup:" destination="-1" eventType="touchUpInside" id="ROG-kC-NZA"/>
+                            </connections>
+                        </button>
+                        <imageView userInteractionEnabled="NO" contentMode="scaleAspectFit" horizontalHuggingPriority="251" verticalHuggingPriority="251" translatesAutoresizingMaskIntoConstraints="NO" id="gmN-gy-bjy">
+                            <rect key="frame" x="110" y="20" width="169" height="126"/>
+                            <constraints>
+                                <constraint firstAttribute="width" constant="169" id="Y9s-pZ-2pq"/>
+                            </constraints>
+                        </imageView>
+                        <label opaque="NO" clipsSubviews="YES" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" text="Message Label" textAlignment="center" lineBreakMode="tailTruncation" numberOfLines="2" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" preferredMaxLayoutWidth="243" translatesAutoresizingMaskIntoConstraints="NO" id="zzP-bh-DHM">
+                            <rect key="frame" x="73" y="199" width="243" height="110"/>
+                            <constraints>
+                                <constraint firstAttribute="height" constant="110" id="dIE-P4-CkO"/>
+                            </constraints>
+                            <fontDescription key="fontDescription" type="boldSystem" pointSize="17"/>
+                            <color key="textColor" red="0.0" green="0.47843137250000001" blue="1" alpha="1" colorSpace="calibratedRGB"/>
+                            <nil key="highlightedColor"/>
+                        </label>
+                    </subviews>
+                    <color key="backgroundColor" white="1" alpha="1" colorSpace="custom" customColorSpace="calibratedWhite"/>
+                    <constraints>
+                        <constraint firstItem="gmN-gy-bjy" firstAttribute="top" secondItem="cYT-D9-Z5J" secondAttribute="top" constant="20" symbolic="YES" id="Jfb-OE-esZ"/>
+                        <constraint firstItem="gmN-gy-bjy" firstAttribute="centerX" secondItem="zzP-bh-DHM" secondAttribute="centerX" id="hBO-UX-EXa"/>
+                        <constraint firstAttribute="bottom" secondItem="Ple-bp-UYe" secondAttribute="bottom" constant="40" id="jam-1d-aCo"/>
+                        <constraint firstItem="zzP-bh-DHM" firstAttribute="leading" secondItem="cYT-D9-Z5J" secondAttribute="leading" constant="73" id="k0E-8Q-Gka"/>
+                        <constraint firstItem="Ple-bp-UYe" firstAttribute="top" secondItem="zzP-bh-DHM" secondAttribute="bottom" constant="70" id="pWB-QQ-55j"/>
+                        <constraint firstItem="Ple-bp-UYe" firstAttribute="centerX" secondItem="cYT-D9-Z5J" secondAttribute="centerX" id="pWk-a1-Nos"/>
+                        <constraint firstItem="zzP-bh-DHM" firstAttribute="centerX" secondItem="Ple-bp-UYe" secondAttribute="centerX" id="wJc-lN-k2Z"/>
+                        <constraint firstItem="zzP-bh-DHM" firstAttribute="top" secondItem="gmN-gy-bjy" secondAttribute="bottom" constant="53" id="xd2-Fl-YNT"/>
+                    </constraints>
+                </view>
+            </subviews>
+            <color key="backgroundColor" white="1" alpha="1" colorSpace="custom" customColorSpace="calibratedWhite"/>
+            <constraints>
+                <constraint firstItem="cYT-D9-Z5J" firstAttribute="leading" secondItem="1" secondAttribute="leading" constant="190" id="7BA-0c-iJw"/>
+                <constraint firstItem="cYT-D9-Z5J" firstAttribute="centerX" secondItem="1" secondAttribute="centerX" id="7OE-dO-zCC"/>
+                <constraint firstAttribute="bottom" secondItem="cYT-D9-Z5J" secondAttribute="bottom" constant="288" id="8u8-93-sv0"/>
+                <constraint firstItem="cYT-D9-Z5J" firstAttribute="top" secondItem="1" secondAttribute="top" constant="287" id="zDT-Rg-ixk"/>
+            </constraints>
+            <nil key="simulatedStatusBarMetrics"/>
+        </view>
+    </objects>
+</document>
