--- NMPopUpView/PopUpViewController.h
@@ -1,21 +1,0 @@
-//
-//  PopUpViewController.h
-//  NMPopUpView
-//
-//  Created by Nikos Maounis on 9/12/13.
-//  Copyright (c) 2013 Nikos Maounis. All rights reserved.
-//
-
-#import <UIKit/UIKit.h>
-#import <QuartzCore/QuartzCore.h>
-#import "AppDelegate.h"
-
-@interface PopUpViewController : UIViewController
-
-@property (weak, nonatomic) IBOutlet UIView *popUpView;
-@property (weak, nonatomic) IBOutlet UIImageView *logoImg;
-@property (weak, nonatomic) IBOutlet UILabel *messageLabel;
-
-- (void)showInView:(UIView *)aView withImage:(UIImage *)image withMessage:(NSString *)message animated:(BOOL)animated;
-
-@end
