--- NMPopUpView/PopUpViewController.m
@@ -1,83 +1,0 @@
-//
-//  PopUpViewController.m
-//  NMPopUpView
-//
-//  Created by Nikos Maounis on 9/12/13.
-//  Copyright (c) 2013 Nikos Maounis. All rights reserved.
-//
-
-#import "PopUpViewController.h"
-
-@interface PopUpViewController ()
-
-@end
-
-@implementation PopUpViewController
-
-- (id)initWithNibName:(NSString *)nibNameOrNil bundle:(NSBundle *)nibBundleOrNil
-{
-    self = [super initWithNibName:nibNameOrNil bundle:nibBundleOrNil];
-    
-    if (self) {
-        // Custom initialization
-    }
-    return self;
-}
-
-- (void)viewDidLoad
-{
-    self.view.backgroundColor=[[UIColor blackColor] colorWithAlphaComponent:.6];
-    self.popUpView.layer.cornerRadius = 5;
-    self.popUpView.layer.shadowOpacity = 0.8;
-    self.popUpView.layer.shadowOffset = CGSizeMake(0.0f, 0.0f);
-    [super viewDidLoad];
-    // Do any additional setup after loading the view from its nib.
-}
-
-- (void)showAnimate
-{
-    self.view.transform = CGAffineTransformMakeScale(1.3, 1.3);
-    self.view.alpha = 0;
-    [UIView animateWithDuration:.25 animations:^{
-        self.view.alpha = 1;
-        self.view.transform = CGAffineTransformMakeScale(1, 1);
-    }];
-    
-}
-
-- (IBAction)closePopup:(id)sender {
-    [self removeAnimate];
-}
-
-- (void)removeAnimate
-{
-    [UIView animateWithDuration:.25 animations:^{
-        self.view.transform = CGAffineTransformMakeScale(1.3, 1.3);
-        self.view.alpha = 0.0;
-    } completion:^(BOOL finished) {
-        if (finished) {
-            [self.view removeFromSuperview];
-        }
-    }];
-}
-
-- (void)showInView:(UIView *)aView withImage:(UIImage *)image withMessage:(NSString *)message animated:(BOOL)animated
-{
-    dispatch_async(dispatch_get_main_queue(), ^{
-    [aView addSubview:self.view];
-    self.logoImg.image = image;
-    self.messageLabel.text = message;
-    if (animated) {
-        [self showAnimate];
-    }
-    });
-}
-
-
-- (void)didReceiveMemoryWarning
-{
-    [super didReceiveMemoryWarning];
-    // Dispose of any resources that can be recreated.
-}
-
-@end
