+++ NMPopUpView/README.md
@@ -1,0 +1,29 @@
+# NMPopUpViewSwift
+
+[![CI Status](http://img.shields.io/travis/Nikos Maounis/NMPopUpViewSwift.svg?style=flat)](https://travis-ci.org/Nikos Maounis/NMPopUpViewSwift)
+[![Version](https://img.shields.io/cocoapods/v/NMPopUpViewSwift.svg?style=flat)](http://cocoapods.org/pods/NMPopUpViewSwift)
+[![License](https://img.shields.io/cocoapods/l/NMPopUpViewSwift.svg?style=flat)](http://cocoapods.org/pods/NMPopUpViewSwift)
+[![Platform](https://img.shields.io/cocoapods/p/NMPopUpViewSwift.svg?style=flat)](http://cocoapods.org/pods/NMPopUpViewSwift)
+
+## Usage
+
+To run the example project, clone the repo, and run `pod install` from the Example directory first.
+
+## Requirements
+
+## Installation
+
+NMPopUpViewSwift is available through [CocoaPods](http://cocoapods.org). To install
+it, simply add the following line to your Podfile:
+
+```ruby
+pod "NMPopUpViewSwift"
+```
+
+## Author
+
+Nikos Maounis, nikospsy2k@gmail.com
+
+## License
+
+NMPopUpViewSwift is available under the MIT license. See the LICENSE file for more info.
