--- NMPopUpView/main.m
@@ -1,18 +1,0 @@
-//
-//  main.m
-//  NMPopUpView
-//
-//  Created by Nikos Maounis on 9/12/13.
-//  Copyright (c) 2013 Nikos Maounis. All rights reserved.
-//
-
-#import <UIKit/UIKit.h>
-
-#import "AppDelegate.h"
-
-int main(int argc, char * argv[])
-{
-    @autoreleasepool {
-        return UIApplicationMain(argc, argv, nil, NSStringFromClass([AppDelegate class]));
-    }
-}
