--- NMPopUpViewTests/NMPopUpViewTests.m
@@ -1,34 +1,0 @@
-//
-//  NMPopUpViewTests.m
-//  NMPopUpViewTests
-//
-//  Created by Nikos Maounis on 9/12/13.
-//  Copyright (c) 2013 Nikos Maounis. All rights reserved.
-//
-
-#import <XCTest/XCTest.h>
-
-@interface NMPopUpViewTests : XCTestCase
-
-@end
-
-@implementation NMPopUpViewTests
-
-- (void)setUp
-{
-    [super setUp];
-    // Put setup code here. This method is called before the invocation of each test method in the class.
-}
-
-- (void)tearDown
-{
-    // Put teardown code here. This method is called after the invocation of each test method in the class.
-    [super tearDown];
-}
-
-- (void)testExample
-{
-    XCTFail(@"No implementation for \"%s\"", __PRETTY_FUNCTION__);
-}
-
-@end
